% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{add_durations}
\alias{add_durations}
\title{Add duration variables}
\usage{
add_durations(
  data,
  exposure_end_variable,
  exposure_start_variable,
  ae_start_variable
)
}
\arguments{
\item{data}{Combined ADAE-ADSL data}

\item{exposure_end_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_start_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{ae_start_variable}{Variables for calculation of time at risk for
incidence rates.}
}
\description{
Add duration variables
}
