% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap_fct_plots.R
\name{add_heatmap_colors}
\alias{add_heatmap_colors}
\title{Color palette for heatmap plot}
\usage{
add_heatmap_colors(
  data,
  heatmap_mode = c("RR", "RD", "FDR", "DFDR"),
  alpha,
  theme = c("light", "dark")
)
}
\arguments{
\item{data}{A dataframe with heatmap data.}

\item{heatmap_mode}{A character, can be "RR", "RD", "FDR", "DFDR".}

\item{alpha}{Significance level.}

\item{theme}{Either "light" or "dark".}
}
\description{
Color palette for heatmap plot
}
