% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap_fct_plots.R
\name{add_heatmap_hover_text}
\alias{add_heatmap_hover_text}
\title{Add hover text to heatmap data}
\usage{
add_heatmap_hover_text(
  data,
  heatmap_color,
  verum_name = "Verum",
  comparator_name = "Comparator"
)
}
\arguments{
\item{data}{Dataframe from make_treemap_data()}

\item{heatmap_color}{Effect measure: "RR", "RD", "FDR" or "DFDR".}

\item{verum_name}{Custom name for Verum arm.}

\item{comparator_name}{Custom name for Comparator arm.}
}
\description{
Add hover text to heatmap data
}
