% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_fct_plots.R
\name{add_hover_text}
\alias{add_hover_text}
\title{Helper function to add hover text to the effect plot}
\usage{
add_hover_text(data, display = c("RD", "RR"), adjustment = c("FDR", "DFDR"))
}
\arguments{
\item{data}{A dataframe.}

\item{display}{Either "RD" or "RR".}

\item{adjustment}{"FDR" or "DFDR"}
}
\description{
Helper function to add hover text to the effect plot
}
