% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{add_incidence_numerators}
\alias{add_incidence_numerators}
\title{Calculate numerators of incidence rates}
\usage{
add_incidence_numerators(
  data,
  variable,
  ae_duration_variable,
  exposure_duration_variable,
  overall = FALSE
)
}
\arguments{
\item{data}{ADAE-ADSL data with duration variables}

\item{variable}{A character, the variable name that will be shown in the plot
("AEDECOD", "MLG_label", "SMQ_NAME", "ocmq" or "AEBODSYS").}

\item{ae_duration_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_duration_variable}{Variables for calculation of time at risk
for incidence rates.}

\item{overall}{Whether to calculate events per AE category, or overall}
}
\description{
Calculate numerators of incidence rates
}
