% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{add_incidence_rates}
\alias{add_incidence_rates}
\title{Calculate incidence rates}
\usage{
add_incidence_rates(data, variable)
}
\arguments{
\item{data}{Combined ADSL-ADAE data with incidence rate numerators and denominators}

\item{variable}{A character, the variable name that will be shown in the plot
("AEDECOD", "MLG_label", "SMQ_NAME", "ocmq" or "AEBODSYS").}
}
\description{
Calculate incidence rates
}
