% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{add_proportions}
\alias{add_proportions}
\title{Add proportions}
\usage{
add_proportions(data)
}
\arguments{
\item{data}{Data with events ("count") and denominators ("big_n")}
}
\description{
Add proportions
}
