% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{arrange_data}
\alias{arrange_data}
\title{A function to prepare data for plotting the double dot plots}
\usage{
arrange_data(
  data,
  order_by = c("p-value", "effect"),
  adjustment = c("FDR", "DFDR"),
  effect_measure = c("RD", "RR"),
  number_aes
)
}
\arguments{
\item{data}{Result data set from function calculate_results.}

\item{order_by}{p-value or effect.}

\item{adjustment}{FDR (False Discovery Rate) or DFDR (Double False Discovery
Rate).}

\item{effect_measure}{RR (Relative Risk) or RD (Risk Difference).}

\item{number_aes}{Number of Adverse Events shown.}
}
\value{
An data object arranged and filtered, ready to plot
}
\description{
Sorts data by order of effect or significance.
}
