% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{assert_columns}
\alias{assert_columns}
\title{Custom assert that required columns are present in data}
\usage{
assert_columns(data, req_columns)
}
\arguments{
\item{data}{A dataframe.}

\item{req_columns}{A character vector of column names required.}
}
\description{
Custom assert that required columns are present in data
}
