% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{assert_is_not_null}
\alias{assert_is_not_null}
\title{Assert that an object is not null}
\usage{
assert_is_not_null(object)
}
\arguments{
\item{object}{The object.}
}
\description{
Assert that an object is not null
}
