% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_info_fct_info.R
\name{calc_summary_data}
\alias{calc_summary_data}
\title{Calculate summary data to display in Dataset Information panel}
\usage{
calc_summary_data(data, data_filtered)
}
\arguments{
\item{data}{An unfiltered dataframe.}

\item{data_filtered}{A filtered dataframe.}
}
\description{
Calculate summary data to display in Dataset Information panel
}
