% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_calcDoubleDotPlot.R
\name{calcDoubleDotPlot2}
\alias{calcDoubleDotPlot2}
\title{Calculate all Variables for the double_dot_plot function}
\usage{
calcDoubleDotPlot2(
  data_adsl,
  data_adae,
  variable = "AEDECOD",
  display = "RR",
  adjustment = "DFDR",
  order_by = "p-value",
  study_strat = "None",
  n.iterations = 1e+05,
  n.burn.in = 5000,
  seed = 2006,
  alternative = "two.sided",
  heatmap = FALSE,
  alpha = 0.05,
  filter = "nomethod",
  measure = "proportions"
)
}
\arguments{
\item{data_adsl}{Subject level data set.}

\item{data_adae}{Adverse event data set.}

\item{variable}{AEDECOD or AEBODSYS or MLG.}

\item{display}{Display option RR (relative risks) or RD (risk differences).}

\item{adjustment}{Adjustment variable FDR or DFDR or BB.}

\item{order_by}{Order of adverse events shown in double dot plot (p-value or effect).}

\item{study_strat}{If study stratification should be used (TRUE or FALSE).}

\item{n.iterations}{Number of iterations for BB.}

\item{n.burn.in}{Number of burn in phase for BB.f}

\item{seed}{Random Number Generator number.}

\item{alternative}{A character string describing the alternative hypotesis (one or two sided)}

\item{heatmap}{Transform returned data into structure of the heatmap}

\item{alpha}{Alpha level of the test}

\item{filter}{A character string if filter on adverse events should be performed}

\item{measure}{Measurement used (proportions or incidence rates)}
}
\value{
results2_data
}
\description{
Calculate all Variables for the double_dot_plot function
}
