% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_adjustment.R
\name{calculate_stratified_incidence}
\alias{calculate_stratified_incidence}
\title{Calculate the study-size adjusted incidence rates}
\usage{
calculate_stratified_incidence(
  split_data,
  effect_measure = c("RD", "RR"),
  level,
  stratify_variable
)
}
\arguments{
\item{split_data}{Data set with Counts.}

\item{effect_measure}{Effect measure, either "RR" or "RD".}

\item{level}{Significance level (e.g., 95).}

\item{stratify_variable}{Stratification variable.}
}
\value{
A one-row tibble with weighted proportions, relative risk,
risk differences, confidence intervals and p-values.
}
\description{
Calculate the study-size adjusted incidence rates
}
