% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_adae_adsl.R
\name{check_adae_data}
\alias{check_adae_data}
\title{Check that adverse event dataset has the required variables}
\usage{
check_adae_data(adae)
}
\arguments{
\item{adae}{A dataframe with ADAE data.}
}
\description{
This function runs after uploading an ADAE file in a valid extension
}
