% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_adae_adsl.R
\name{check_adsl_data}
\alias{check_adsl_data}
\title{Check that ADSL dataset has the required variables}
\usage{
check_adsl_data(adsl)
}
\arguments{
\item{adsl}{A dataframe with ADSL data.}
}
\description{
This function runs after uploading an ADSL file in a valid extension
}
