% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_meddra.R
\name{check_meddra_data}
\alias{check_meddra_data}
\title{Check MedDRA data from uploaded files}
\usage{
check_meddra_data(data)
}
\arguments{
\item{data}{A dataframe from uploaded files. Must have columns MT_PT,
MT_HLT, MT_HLGT, VERSION (all characters).}
}
\value{
A tibble with necessary MedDRA variables.
}
\description{
Check MedDRA data from uploaded files
}
