% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_meddra.R
\name{check_smq_data}
\alias{check_smq_data}
\title{Check MedDRA-SMQ data from uploaded files}
\usage{
check_smq_data(data)
}
\arguments{
\item{data}{A dataframe from uploaded files.}
}
\value{
A tibble with necessary MedDRA-SMQ variables.
}
\description{
Check MedDRA-SMQ data from uploaded files
}
