% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_volcano.R
\name{color_volcano}
\alias{color_volcano}
\title{Color volcano circles according to significance and effect}
\usage{
color_volcano(data, effect_measure, alpha)
}
\arguments{
\item{data}{Data prepared for volcano plot.}

\item{effect_measure}{Either "RR" or "RD".}

\item{alpha}{Significance level.}
}
\description{
Color volcano circles according to significance and effect
}
