% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_heatmap.R
\name{count_events_by_arm}
\alias{count_events_by_arm}
\title{Count total adverse events by safety variable and treatment arm}
\usage{
count_events_by_arm(data, variable)
}
\arguments{
\item{data}{A dataframe of adverse events, with columns: a safety variable
(AEBODSYS, MT_HLT, MT_HLGT, or SOC_MLG), "count", and
"trta_detector" (treatment arm)}

\item{variable}{A character, the safety variable to group the count}
}
\value{
A wide dataframe with values: counts, rows: AE categories,
columns: treatment arms
}
\description{
Count total adverse events by safety variable and treatment arm
}
