% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_heatmap.R
\name{count_heatmap_data}
\alias{count_heatmap_data}
\title{Count total counts and ratios Comparator/Verum for heatmap map}
\usage{
count_heatmap_data(
  data,
  heatmap_variable = c("AEDECOD", "AEDECOD_HLT", "MLG_label")
)
}
\arguments{
\item{data}{A dataframe after prepare_heatmap_data()}

\item{heatmap_variable}{A character, the safety hierarchy variable. Can be
"AEDECOD", "AEDECOD_HLT" or "MLG_label"}
}
\value{
A dataframe of counts of events for each treatment arm
(and ratio) in each hierarchy category
}
\description{
Creates the total counts of events in Comparator and Verum group in each of
the MedDRA hierarchy categories.
}
