% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_info_fct_info.R
\name{count_subjects}
\alias{count_subjects}
\title{Helper function to calculate the count of unique USUBJID by treatment arm}
\usage{
count_subjects(data, arm = c("Verum", "Comparator"), filter = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{arm}{A treatment arm, can be "Verum" or "Comparator".}

\item{filter}{A variable name to filter not null values.}
}
\description{
Helper function to calculate the count of unique USUBJID by treatment arm
}
