% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_adjustment.R
\name{count_total_proportions}
\alias{count_total_proportions}
\title{Count proportions of total AEs by AE category}
\usage{
count_total_proportions(data, variable)
}
\arguments{
\item{data}{A dataframe, with columns "variable" (AEDECOD or MLG_label),
"count", "big_n" and "p"}

\item{variable}{A character, either "AEDECOD" or "MLG_label"}
}
\value{
A dataframe with total Count of events, and proportion of all events
(column "prop")
}
\description{
Count proportions of total AEs by AE category
}
