% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap_fct_plots.R
\name{create_heatmap_palette}
\alias{create_heatmap_palette}
\title{Create a discrete color palette for heatmap}
\usage{
create_heatmap_palette(theme = c("light", "dark"))
}
\arguments{
\item{theme}{Either "light" or "dark"}
}
\description{
Create a discrete color palette for heatmap
}
