% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectoR-package.R
\docType{package}
\name{DetectoR-package}
\alias{DetectoR}
\alias{DetectoR-package}
\title{DetectoR: Stay Safe; Don't Miss Any Adverse-Event Signal}
\description{
The DetectoR R Shiny app allows a rapid, rigorous exploration of safety data to uncover hidden adverse‑event signals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bayer-group/DetectoR}
  \item Report bugs at \url{https://github.com/bayer-group/DetectoR/issues}
}

}
\author{
\strong{Maintainer}: Carlos Fernandez-Escobar \email{carlos.fernandez4.ext@bayer.com}

Authors:
\itemize{
  \item Steffen Jeske
  \item Martin Gebel
  \item Ann-Kathrin Frenz
}

}
\keyword{internal}
