% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{detector_pretty_toggle}
\alias{detector_pretty_toggle}
\title{Default shinyWidgets::prettyToggle()}
\usage{
detector_pretty_toggle(id, label, label_off = NULL, icon, icon_off = NULL, ...)
}
\arguments{
\item{id}{A character, input's id.}

\item{label}{A character, label to show on input.}

\item{label_off}{(Optional) A character, label to show on input when off.}

\item{icon}{A character, an icon to be passed to shiny::icon().}

\item{icon_off}{(Optional) A character, an icon to be passed to shiny::icon()
when off.}

\item{...}{Optional additional parameters to prettyToggle().}
}
\description{
Default shinyWidgets::prettyToggle()
}
