% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{detector_radio_buttons}
\alias{detector_radio_buttons}
\title{Default prettyRadioButtons}
\usage{
detector_radio_buttons(input_id, label, ...)
}
\arguments{
\item{input_id}{Input ID.}

\item{label}{Text to show in the input.}

\item{...}{Options to pass to shinyWidgets::prettyRadioButtons}
}
\description{
Default prettyRadioButtons
}
