% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_info_fct_info.R
\name{display_info_summary}
\alias{display_info_summary}
\title{Display summary info}
\usage{
display_info_summary(
  data_unfiltered,
  data_filtered,
  adae_unfiltered,
  adae_filtered,
  verum_name,
  comparator_name
)
}
\arguments{
\item{data_unfiltered}{Joined data before applying filters.}

\item{data_filtered}{Joined data after applying filters.}

\item{adae_unfiltered}{Full ADAE.}

\item{adae_filtered}{ADAE after applying filters.}

\item{verum_name}{Custom verum arm name.}

\item{comparator_name}{Custom comparator arm name.}
}
\description{
Display summary info
}
