% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_graphing.R
\name{double_dot_plot_legend}
\alias{double_dot_plot_legend}
\title{A function to plot the legend of the double dot plot}
\usage{
double_dot_plot_legend(
  data = results2_data,
  adjustment,
  variable,
  display,
  numberAEs = 50,
  color_theme = "dark"
)
}
\arguments{
\item{data}{Data set.}

\item{adjustment}{If adjustment should be peformed ("DFDR"/"FDR")}

\item{variable}{A character string with the variable name ("AEBODSYS"/"AEDECOD")}

\item{display}{Order to display.}

\item{numberAEs}{Number of adverse events shown.}

\item{color_theme}{Theme for the legend.}
}
\value{
An object with legend graphics
}
\description{
A function to plot the legend of the double dot plot
}
