% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_graphing.R
\name{double_dot_plot_p1}
\alias{double_dot_plot_p1}
\title{A function to plot the double_dot_plot / Relative risk / adjusted p-values and p-values}
\usage{
double_dot_plot_p1(
  data = results2_data,
  adjustment,
  variable,
  display,
  numberAEs = 50,
  label = PT,
  color_theme = "dark",
  measure = "proportions"
)
}
\arguments{
\item{data}{Result data set from function calcDoubleDotPlot.}

\item{adjustment}{FDR or DFDR or BB.}

\item{variable}{AEDECOD or AEBODSYS.}

\item{display}{RR or RD.}

\item{numberAEs}{Number of Adverse Events shown (default = 50, have to be a even number)}

\item{label}{Label variable.}

\item{color_theme}{Select color theme.}

\item{measure}{Measure variable (proportions or incidence rates)}
}
\value{
An object with 4 ggplot graphics
}
\description{
A function to plot the double_dot_plot / Relative risk / adjusted p-values and p-values
}
