% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap_fct_plots.R
\name{draw_treemap}
\alias{draw_treemap}
\title{Draw a treemap plot}
\usage{
draw_treemap(data, heatmap_color, theme = c("light", "dark"))
}
\arguments{
\item{data}{Data prepared for treemap.}

\item{heatmap_color}{"RD", "RR", "FDR" or "DFDR".}

\item{theme}{Either "light" or "dark".}
}
\description{
Draw a treemap plot
}
