% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_volcano_fct_plots.R
\name{draw_volcano}
\alias{draw_volcano}
\title{Draw volcano plot}
\usage{
draw_volcano(data, effect_measure, verum_name, comparator_name, alpha)
}
\arguments{
\item{data}{A dataframe.}

\item{effect_measure}{Either "RR" or "RD".}

\item{verum_name}{Custom verum name.}

\item{comparator_name}{Custom comparator name.}

\item{alpha}{Significance level.}
}
\description{
Draw volcano plot
}
