% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{FDR.fun}
\alias{FDR.fun}
\title{Calculate the false discovery rate}
\usage{
FDR.fun(ADAE)
}
\arguments{
\item{ADAE}{The adae data set.}
}
\value{
ordered.data
}
\description{
Calculate the false discovery rate
}
