% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{filter_ae_type}
\alias{filter_ae_type}
\title{Filter by AE type (Serious, Treatment-Emergent, etc.)}
\usage{
filter_ae_type(
  data,
  type = c("treatment_emergent", "serious_treatment_emergent",
    "drug_related_treatment_emergent", "serious_drug_related_treatment_emergent", "all",
    "serious", "drug_related", "serious_drug_related"),
  treatment_emergent_variable,
  treatment_emergent_value,
  serious_variable,
  serious_value,
  drug_related_variable,
  drug_related_value
)
}
\arguments{
\item{data}{Data to filter on.}

\item{type}{AE type.}

\item{treatment_emergent_variable}{TEAE flag variable, e.g. "TRETMFL".}

\item{treatment_emergent_value}{TEAE flag value, e.g. "Y".}

\item{serious_variable}{SAE flag variable, e.g. "AESER".}

\item{serious_value}{SAE flag value, e.g. "Y".}

\item{drug_related_variable}{Drug-related AE flag variable, e.g. "AEREL".}

\item{drug_related_value}{Drug-related AE flag value, e.g. "Y".}
}
\description{
Filter by AE type (Serious, Treatment-Emergent, etc.)
}
