% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{filter_empty_variable}
\alias{filter_empty_variable}
\title{Helper for filtering out empty values and convert to factor}
\usage{
filter_empty_variable(data, variable)
}
\arguments{
\item{data}{A dataframe}

\item{variable}{A character variable}
}
\description{
Helper for filtering out empty values and convert to factor
}
