% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{filter_minimum_aes}
\alias{filter_minimum_aes}
\title{Filter the minimum count of AEs}
\usage{
filter_minimum_aes(
  data,
  variable,
  big_n,
  method = c("no_method", "one_percent", "min"),
  alternative = c("two.sided", "less", "greater"),
  alpha
)
}
\arguments{
\item{data}{A dataframe, the combined data of AEs.}

\item{variable}{A character, the safety variable used to group AEs.}

\item{big_n}{A dataframe, the summarised data with number of subjects per
treatment arm.}

\item{method}{A character, the method of filtering: "no_method",
"one_percent" or "min". "min" calls find_min_event()}

\item{alternative}{A character describing the alternative hypothesis for
"min" method: "two.sided", "less" or "greater" (one-sided)}

\item{alpha}{Numeric, alpha level of the test for "min" method.}
}
\value{
A dataframe of filtered adverse events.
}
\description{
Takes a dataframe of adverse events (AEs) and filters it
 according to a minimum count criteria: "one_percent" (incidence proportion
 >=1\%) or "min" (minimum number of events in Verum arm).
}
