% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{find.min.event}
\alias{find.min.event}
\title{Calculate the minimum number of events in verum}
\usage{
find.min.event(N1, N2, alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{N1}{Number of patients in Verum}

\item{N2}{Number of patients in Comparator}

\item{alternative}{A character string describing the alternative hypotesis (one or two sided)}

\item{alpha}{Alpha level of the test}
}
\value{
minimun number of events in verum
}
\description{
Calculate the minimum number of events in verum
}
