% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fisher.and.rr}
\alias{fisher.and.rr}
\title{Perform fisher test and calculate the relative risk and risk difference (Version 2)}
\usage{
fisher.and.rr(ds, alternative = "two.sided", alph = 0.05, vari)
}
\arguments{
\item{ds}{Data set.}

\item{alternative}{Either one or two sided test.}

\item{alph}{Significance level.}

\item{vari}{Target column.}
}
\value{
output A tibble with the results.
}
\description{
Perform fisher test and calculate the relative risk and risk difference (Version 2)
}
