% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_fisher.and.rr.R
\name{fisher.and.rr_ir}
\alias{fisher.and.rr_ir}
\title{Perform fisher test and calculate the relative risk and risk difference based on incidence rates}
\usage{
fisher.and.rr_ir(ds, alternative = "two.sided", alph = 0.05, vari)
}
\arguments{
\item{ds}{Data set.}

\item{alternative}{Either one or two sided test.}

\item{alph}{Significance level.}

\item{vari}{Target column.}
}
\value{
output A tibble with the results.
}
\description{
Perform fisher test and calculate the relative risk and risk difference based on incidence rates
}
