% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_fct_plots.R
\name{flag_significant}
\alias{flag_significant}
\title{Flags p-values by statistical significance}
\usage{
flag_significant(
  data,
  adjustment = c("FDR", "DFDR"),
  effect_measure = c("RD", "RR"),
  alpha
)
}
\arguments{
\item{data}{A dataframe of results. Must include columns called "p" and
"p_adj"}

\item{adjustment}{A character, "FDR" or "DFDR"}

\item{effect_measure}{Either "RR" or "RD"}

\item{alpha}{Numeric, the significance level}
}
\value{
A dataframe with colors for p-values in "col_p", and for adjusted
p-values in "col_p_adj"
}
\description{
Flags p-values by statistical significance
}
