% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{format_decimals}
\alias{format_decimals}
\title{Format number to decimal places}
\usage{
format_decimals(number, digits)
}
\arguments{
\item{number}{A numeric vector}

\item{digits}{Number of decimal places}
}
\description{
Format number to decimal places
}
