% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{format_p_values}
\alias{format_p_values}
\title{Format p-values to four decimal places}
\usage{
format_p_values(p)
}
\arguments{
\item{p}{A numeric vector of p-values}
}
\value{
A character vector of p-values rounded to four decimal places, or
"<0.0001"
}
\description{
Format p-values to four decimal places
}
