% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{get_big_n}
\alias{get_big_n}
\title{Get the denominator counts (N for arms) for proportions or incidence rates}
\usage{
get_big_n(
  adsl_filtered_data,
  frequency_measure = c("proportions", "incidence rates"),
  duration_mode,
  exposure_duration_variable = NULL,
  exposure_start_variable = NULL,
  exposure_end_variable = NULL
)
}
\arguments{
\item{adsl_filtered_data}{ADSL with filters applied}

\item{frequency_measure}{Either "proportions" or "incidence rates"}

\item{duration_mode}{Way of calculating incidence risks time at risk,
can be "duration", "start_end_date" or "none".}

\item{exposure_duration_variable}{Variables for calculation of time at risk
for incidence rates.}

\item{exposure_start_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_end_variable}{Variables for calculation of time at risk for
incidence rates.}
}
\value{
A tibble with columns trta_detector, big_n (number of subjects),
big_exp (total duration of exposure, only if incidence rates)
}
\description{
Get the denominator counts (N for arms) for proportions or incidence rates
}
