% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_heatmap.R
\name{get_child_parent}
\alias{get_child_parent}
\title{Get equivalence of children-parents hierarchical categories}
\usage{
get_child_parent(data, child, parent)
}
\arguments{
\item{data}{A dataframe with a child variable and parent variable}

\item{child}{A character, the name of the child variable}

\item{parent}{A character, the name of the parent variable. If NULL,
the parent column will be an empty string}
}
\value{
A data frame with two columns: "child" and "parent"
}
\description{
From a table, it extracts unique combinations of child-parent variables
}
