% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{get_count_proportions}
\alias{get_count_proportions}
\title{Calculate counts and proportions of events}
\usage{
get_count_proportions(
  comb_data,
  big_n,
  variable = c("AEDECOD", "MLG_label", "SMQ_NAME", "AEBODSYS"),
  frequency_measure = c("proportions", "incidence rates"),
  duration_mode,
  ae_duration_variable = NULL,
  exposure_duration_variable = NULL,
  exposure_start_variable = NULL,
  exposure_end_variable = NULL,
  ae_start_variable = NULL
)
}
\arguments{
\item{comb_data}{Filtered dataset combining ADSL and ADAE}

\item{big_n}{Tibble with denominators for treatment arms (number of subjects)}

\item{variable}{A character, the variable name that will be shown in the plot
("AEDECOD", "MLG_label", "SMQ_NAME" or "AEBODSYS").}

\item{frequency_measure}{Either "proportions" or "incidence rates"}

\item{duration_mode}{Way of calculating incidence risks denominators (time at
risk),
can be "duration", "start_end_date" or "none".}

\item{ae_duration_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_duration_variable}{Variables for calculation of time at risk
for incidence rates.}

\item{exposure_start_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_end_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{ae_start_variable}{Variables for calculation of time at risk for
incidence rates.}
}
\value{
A tibble with columns:
\describe{
  \item{\code{variable}}{The safety variable, can be "AEDECOD", "MLG_label",
"SMQ_NAME" or "AEBODSYS"}
  \item{\code{trta_detector}}{Treatment arm, "Verum" or "Comparison"}
  \item{\code{count}}{Number of events}
  \item{\code{no_count}}{Number of no-events (total minus number of events)}
  \item{\code{big_n}}{Number of subjects in each arm}
  \item{\code{prop}}{Proportion or incidence rate}
  \item{\code{pattime}}{Text exposure time per arm, only if incidence rates}
}
}
\description{
Calculate counts and proportions of events
}
