% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_exponent}
\alias{get_exponent}
\title{Purpose: get exponent of the mean of a variable in scientific notation
         for adaptive rounding
         When abs(mean)>=1, 0 is returned}
\usage{
get_exponent(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
integer
}
\description{
Purpose: get exponent of the mean of a variable in scientific notation
         for adaptive rounding
         When abs(mean)>=1, 0 is returned
}
