% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_adjustment.R
\name{get_fdr}
\alias{get_fdr}
\title{Calculate the false discovery rate (this is an almost duplicate of
c212.BH.adjust.pvals())}
\usage{
get_fdr(data)
}
\arguments{
\item{data}{The ADAE data set.}
}
\value{
An ordered dataframe with FDR p-values
}
\description{
Calculate the false discovery rate (this is an almost duplicate of
c212.BH.adjust.pvals())
}
