% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_adjustment.R
\name{get_fdr_p_values}
\alias{get_fdr_p_values}
\title{Adjust p-values using Benjamini-Hochberg False Discovery Rate (FDR) method}
\usage{
get_fdr_p_values(data)
}
\arguments{
\item{data}{A dataframe of results. Must include a column named "p", with
unadjusted p-values}
}
\value{
A dataframe with adjusted p-values, in a column named "p_adj"
}
\description{
Adjust p-values using Benjamini-Hochberg False Discovery Rate (FDR) method
}
