% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_effect.R
\name{get_fisher_p_value}
\alias{get_fisher_p_value}
\title{Calculates Fisher's exact test p-values}
\usage{
get_fisher_p_value(a, b, c, d, alternative)
}
\arguments{
\item{a, b, c, d}{Numeric, each one of the variables to compare. In proportions,
 counts of cases and no-cases. In incidence rates, counts of cases and
patient-times.}

\item{alternative}{Character, indicates the alternative hypothesis and must
be "two.sided", "greater" or "less"}
}
\value{
A numeric vector with p-values
}
\description{
Obtains p-values from Fisher's exact test on a 2 x 2 table.
}
