% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_effect.R
\name{get_fisher_rr_rates}
\alias{get_fisher_rr_rates}
\title{Perform fisher test and calculate the relative risk and risk difference based
 on incidence rates}
\usage{
get_fisher_rr_rates(data, alternative = "two.sided", alpha = 0.05, variable)
}
\arguments{
\item{data}{Data set. Must have variables trta_detector, count, no_count,
prop, big_n, pattime}

\item{alternative}{Either one or two sided test.}

\item{alpha}{Significance level.}

\item{variable}{Target column.}
}
\value{
output A tibble with the results.
}
\description{
Perform fisher test and calculate the relative risk and risk difference based
 on incidence rates
}
