% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_adjustment.R
\name{get_new_dfdr}
\alias{get_new_dfdr}
\title{Calculate the new DFDR adjusted p-values}
\usage{
get_new_dfdr(data, variable = c("AEDECOD", "MLG_label"), alpha)
}
\arguments{
\item{data}{A dataframe with "p", lower AE category ("AEDECOD" or
"MLG_label"), and higher AE category ("AEBODSYS" or "SOC_MLG")}

\item{variable}{A character, the lower AE category ("AEDECOD" or "MLG_label")}

\item{alpha}{Numeric, the significance level}
}
\value{
A dataframe with the adjusted p-values in column "DFDR"
}
\description{
Calculate the new DFDR adjusted p-values
}
