% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_heatmap.R
\name{get_parents_heatmap}
\alias{get_parents_heatmap}
\title{Get MedDRA equivalence children-parent category}
\usage{
get_parents_heatmap(
  data,
  heatmap_variable = c("AEDECOD", "AEDECOD_HLT", "MLG_label")
)
}
\arguments{
\item{data}{A dataframe with AE categories. Obtained after
prepare_heatmap_data()}

\item{heatmap_variable}{A character, the safety hierarchy variable. Can be
"AEDECOD", "AEDECOD_HLT" or "MLG_label"}
}
\description{
Obtains a dataframe of equivalence child-parent MedDRA category,
for a particular hierarchy
}
