% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{get_rates_data}
\alias{get_rates_data}
\title{Get rates data}
\usage{
get_rates_data(
  data,
  variable,
  ae_duration_variable,
  exposure_duration_variable,
  overall = FALSE,
  data_denominators
)
}
\arguments{
\item{data}{ADAE-ADSL data with duration variables}

\item{variable}{A character, the variable name that will be shown in the plot
("AEDECOD", "MLG_label", "SMQ_NAME", "ocmq" or "AEBODSYS").}

\item{ae_duration_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_duration_variable}{Variables for calculation of time at risk
for incidence rates.}

\item{overall}{Whether to calculate events per AE category, or overall}

\item{data_denominators}{Count of denominators}
}
\description{
Get rates data
}
