% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_effect.R
\name{get_rd}
\alias{get_rd}
\title{Calculates Risk Difference and Confidence Intervals for proportions or
incidence rates}
\usage{
get_rd(
  x1,
  n1,
  x2,
  n2,
  method = "wald",
  mode = c("proportions", "incidence rates"),
  alpha
)
}
\arguments{
\item{x1}{Numeric, number of cases in verum group}

\item{n1}{Numeric, patient-time at risk (incidence rates) or number of
patients (proportions) in verum group}

\item{x2}{Numeric, number of cases in verum group}

\item{n2}{Numeric, patient-time at risk (incidence rates) or number of
patients (proportions) in comparison group}

\item{method}{Character, by default "wald". Any of the methods of CI
calculation accepted by DescTools::BinomDiffCI()}

\item{mode}{Character, "proportions" or "incidence rates"}

\item{alpha}{Numeric, significance level}
}
\value{
A tibble, with columns named "rd" (estimate), "rd_lcl" (lower CI
limit), and "rd_ucl" (upper CI limit)
}
\description{
Calculates Risk Difference (RD) and Confidence Intervals (CI),
for proportions or incidence rates of Adverse Events. It calls
DescTools::BinomDiffCI().
}
