% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_effect.R
\name{get_rr}
\alias{get_rr}
\title{Calculates Risk Ratio and Confidence Intervals for proportions or incidence
rates}
\usage{
get_rr(
  x1,
  n1,
  x2,
  n2,
  method = "katz.log",
  mode = c("proportions", "incidence rates"),
  alpha
)
}
\arguments{
\item{x1}{Numeric, number of cases in verum group}

\item{n1}{Numeric, patient-time at risk (incidence rates) or number of
patients (proportions) in verum group}

\item{x2}{Numeric, number of cases in verum group}

\item{n2}{Numeric, patient-time at risk (incidence rates) or number of
patients (proportions) in comparison group}

\item{method}{Character, by default "katz.log". Any of the methods of CI
calculation accepted by DescTools::BinomRatioCI()}

\item{mode}{Character, "proportions" or "incidence rates"}

\item{alpha}{Numeric, significance level}
}
\value{
A tibble, with columns named "rr" (estimate), "rr_lcl" (lower CI
limit), and "rr_ucl" (upper CI limit)
}
\description{
Calculates Risk Ratio (RR) and Confidence Intervals (CI), for
proportions or incidence rates of Adverse Events, using a zero-value
correction. It calls DescTools::BinomRatioCI().
}
