% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_adjustment.R
\name{get_stratified_rr_rd}
\alias{get_stratified_rr_rd}
\title{Calculate stratified relative risk and risk differences}
\usage{
get_stratified_rr_rd(
  comb_data,
  variable,
  effect_measure = c("RD", "RR"),
  alpha,
  stratify_variable
)
}
\arguments{
\item{comb_data}{adsl and adae data set combined}

\item{variable}{Safety variable.}

\item{effect_measure}{Effect measure, either "RR" or "RD".}

\item{alpha}{Significance level.}

\item{stratify_variable}{Stratification variable.}
}
\value{
results
}
\description{
Calculate stratified relative risk and risk differences
}
