% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Extract variable names and labels to show as options in pickerInput}
\usage{
get_variable_labels(data)
}
\arguments{
\item{data}{A dataframe (ADAE or ADSL)}
}
\description{
Extract variable names and labels to show as options in pickerInput
}
