% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{invisible_text}
\alias{invisible_text}
\title{Shorthand for invisible text (triggers conditional panels, but it's not
visible)}
\usage{
invisible_text(...)
}
\arguments{
\item{...}{Contents of the row.}
}
\description{
Shorthand for invisible text (triggers conditional panels, but it's not
visible)
}
