% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_join.R
\name{join_adsl_adae}
\alias{join_adsl_adae}
\title{Join ADAE and ADSL datasets}
\usage{
join_adsl_adae(adsl, adae)
}
\arguments{
\item{adsl}{A dataframe, output from `adsl_data_trta()`.}

\item{adae}{A dataframe, output from `select_adae_data()`.}
}
\value{
adsl_adae_data, a dataframe of joined ADAE-ADSL data.
}
\description{
Join ADAE and ADSL datasets
}
