% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_join.R
\name{join_meddra_data}
\alias{join_meddra_data}
\title{Join study data with MedDRA dictionary}
\usage{
join_meddra_data(adsl_adae_data, mode = c("hlt", "mlg", "smq"), meddra_data)
}
\arguments{
\item{adsl_adae_data}{A dataframe, output from `join_adsl_adae()`.}

\item{mode}{A character "hlt" (SOC/HLGTs/HLTs/PTs), "mlg" (SOC/MLGs/PTs) or
"smq" (SMQs).}

\item{meddra_data}{MedDRA data from `prepare_meddra_data()`.}
}
\value{
Transformed adverse event dataset.
}
\description{
Join study data with MedDRA dictionary
}
