% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_ocmq.R
\name{join_ocmq_data}
\alias{join_ocmq_data}
\title{Join study data with OCMQ table}
\usage{
join_ocmq_data(adsl_adae_data)
}
\arguments{
\item{adsl_adae_data}{A dataframe, output from `join_adsl_adae()`.}
}
\description{
Join study data with OCMQ table
}
