% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_fct_plots.R
\name{make_double_dot_plotly}
\alias{make_double_dot_plotly}
\title{A plotly function to plot the double_dot_plot}
\usage{
make_double_dot_plotly(
  data,
  label_collapsed,
  frequency_measure = c("proportions", "incidence rates"),
  height
)
}
\arguments{
\item{data}{Arranged data set from arrange_data().}

\item{label_collapsed}{Logical, using collapsed labels or full label.}

\item{frequency_measure}{Measure variable (proportions or incidence rates).}

\item{height}{Height of the plot in pixels.}
}
\value{
A plotly object
}
\description{
A plotly function to plot the double_dot_plot
}
