% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_heatmap.R
\name{make_treemap_data}
\alias{make_treemap_data}
\title{Create treemap data}
\usage{
make_treemap_data(
  data,
  data_counts,
  heatmap_variable = c("AEDECOD", "AEDECOD_HLT", "MLG_label"),
  heatmap_color = c("RR", "RD", "FDR", "DFDR"),
  alpha
)
}
\arguments{
\item{data}{A data prepared for heatmap, output from prepare_heatmap_data()}

\item{data_counts}{A count data, output from count_heatmap_data()}

\item{heatmap_variable}{A character, the safety hierarchy variable. Can be
"AEDECOD", "AEDECOD_HLT" or "MLG_label".}

\item{heatmap_color}{Effect measure: "RR", "RD", "FDR" or "DFDR".}

\item{alpha}{Significance level.}
}
\description{
Create treemap data
}
