% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap_fct_plots.R
\name{make_treemap_legend}
\alias{make_treemap_legend}
\title{Create legend for heatmap}
\usage{
make_treemap_legend(
  data,
  heatmap_color,
  alpha,
  theme,
  verum_name = "Verum",
  comparator_name = "Comparator"
)
}
\arguments{
\item{data}{Dataframe for treemap.}

\item{heatmap_color}{A character: "RD", "RR", "FDR" or "DFDR".}

\item{alpha}{Numeric. Significance level.}

\item{theme}{Either "light" or "dark".}

\item{verum_name}{Custom Verum name.}

\item{comparator_name}{Custom Comparator name.}
}
\description{
Create legend for heatmap
}
