% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph.R
\name{mod_graph_ui_vars}
\alias{mod_graph_ui_vars}
\title{Reactive values for mod_graph_ui Function module}
\usage{
mod_graph_ui_vars(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\value{
list with following components
\describe{
  \item{go_tab2}{go_tab2}
}
}
\description{
Reactive values for mod_graph_ui Function module
}
