% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap.R
\name{mod_heatmap_ui_vars}
\alias{mod_heatmap_ui_vars}
\title{Reactive values for mod_heatmap_ui Function module}
\usage{
mod_heatmap_ui_vars(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\value{
list with following components
\describe{
  \item{go_tab3}{go_tab3}
}
}
\description{
Reactive values for mod_heatmap_ui Function module
}
