% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_safety.R
\name{mod_safety_ui_vars}
\alias{mod_safety_ui_vars}
\title{Reactive values for mod_safety_ui Function module}
\usage{
mod_safety_ui_vars(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\description{
Reactive values for mod_safety_ui Function module
}
