% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_volcano.R
\name{mod_volcano_server}
\alias{mod_volcano_server}
\title{Volcano plot server function}
\usage{
mod_volcano_server(id, r)
}
\arguments{
\item{id}{Internal parameters for \code{shiny}.}

\item{r}{A \code{shiny::reactiveValues()} list for internal module
communication.
\describe{
  \item{\code{mode}}{Either "demo" (using demo data) or "sas" (using ADAM
files).}
  \item{\code{meddra_mode}}{Either "with_meddra" (using MedDRA data) or
"without_meddra".}
  \item{\code{meddra_data}}{A dataframe of filtered and prepared MedDRA
data.}
  \item{\code{smq_data}}{A dataframe of filtered and prepared MedDRA SMQ
data.}
  \item{\code{mlg_data}}{A dataframe of filtered and prepared MedDRA MLG
data.}
  \item{\code{meddra_version}}{A character with the MedDRA version used,
e.g., "28.1".}
  \item{\code{treatment_variable}}{A character. The variable name of the
treatment arm variable in ADSL dataset, e.g., "TR01TA".}
  \item{\code{verum_name}}{A character. The custom name for the verum arm.}
  \item{\code{comparator_name}}{A character. The custom name for the
comparison arm.}
  \item{\code{adsl_data}}{A dataframe of filtered and prepared ADSL data.}
  \item{\code{adae_data}}{A dataframe of filtered and prepared ADAE data.}
  \item{\code{adsl_variable_names}}{A character vector of variable names in
ADSL. Used for selecting variable filters.}
  \item{\code{unfiltered_data}}{A dataframe of joined ADSL-ADAE data, before
applying variable filters.}
  \item{\code{filtered_data}}{A dataframe of joined ADSL-ADAE data, after
applying variable filters.}
  \item{\code{filter_list}}{A vector describing the filters applied to ADSL
and ADAE data.}
  \item{\code{frequency_measure}}{Either "proportions" or "incidence rates".}
  \item{\code{effect_measure}}{Either "RR" (relative risk) or "RD" (risk
difference).}
  \item{\code{adjustment_method}}{Either "FDR" (False Discovery Rate) or
"DFDR" (New Double False Discovery Rate) method of p-value adjustment for
multiplicity.}
  \item{\code{stratified_by}}{A variable name, if calculations are stratified
 by it. "None" if the results are not stratified.}
  \item{\code{show_full_labels}}{An integer. Showing full or collapsed labels
 for double dot plot.}
  \item{\code{alpha}}{A character. The significance level to calculate
p-values and confidence intervals, e.g., "0.05".}
  \item{\code{results}}{A dataframe ready to be plotted.}
  \item{\code{go_calculate}}{An integer. Updates each time "Calculate!" is
pressed, and updates the calculations and plots.}
  \item{\code{parent_session}}{The session of the app server. Used to control
 the change in tabs, e.g., when pressing "Next page".}
}}
}
\description{
A shiny Module for the volcano plot in DetectoR - server part
}
