% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_volcano.R
\name{mod_volcano_ui}
\alias{mod_volcano_ui}
\title{volcano UI Function}
\usage{
mod_volcano_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny.}
}
\description{
A shiny Module for the volcano plot in DetectoR - user interface
}
