% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_adae_adsl.R
\name{prepare_adae_data}
\alias{prepare_adae_data}
\title{Select adverse event data (used in DetectoR app)}
\usage{
prepare_adae_data(mode = c("demo", "sas"), adae = NULL)
}
\arguments{
\item{mode}{A character, "sas" or "demo". Whether to use uploaded files or
demo data.}

\item{adae}{A dataframe. ADAE dataset output from check_adae_data(). NULL if
using demo data.}
}
\value{
Transformed adverse event dataset.
}
\description{
Select adverse event data (used in DetectoR app)
}
