% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_adae_adsl.R
\name{prepare_adsl_data}
\alias{prepare_adsl_data}
\title{Select subject level data (used in DetectoR app)}
\usage{
prepare_adsl_data(mode = c("demo", "sas"), adsl = NULL)
}
\arguments{
\item{mode}{A character, "sas" or "demo". Whether to use uploaded files or
demo data.}

\item{adsl}{A dataframe. ADSL dataset output from check_adsl_data(). NULL if
using demo data.}
}
\value{
A filtered subject level data set.
}
\description{
Select subject level data (used in DetectoR app)
}
