% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_heatmap.R
\name{prepare_heatmap_data}
\alias{prepare_heatmap_data}
\title{Prepare data for heatmap plot}
\usage{
prepare_heatmap_data(
  data,
  heatmap_variable = c("AEDECOD", "AEDECOD_HLT", "MLG_label"),
  heatmap_color,
  adae_data
)
}
\arguments{
\item{data}{A dataframe obtained from \code{calculate_results()}.}

\item{heatmap_variable}{A character, the safety hierarchy variable. Can be
"AEDECOD", "AEDECOD_HLT" or "MLG_label".}

\item{heatmap_color}{Effect measure: "RR", "RD", "FDR" or "DFDR".}

\item{adae_data}{Original ADAE data, to obtain the PT-SOC equivalence.}
}
\value{
A dataframe with unique combinations of MedDRA hierarchies and
a total count of events in each category
}
\description{
Prepare data for rendering the heatmap plot view. It counts the total number
of events for each category in the lowest hierarchy level.
}
