% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_fct_check_meddra.R
\name{prepare_meddra_data}
\alias{prepare_meddra_data}
\title{Prepare MedDRA data from uploaded files}
\usage{
prepare_meddra_data(data, meddra_version)
}
\arguments{
\item{data}{A dataframe from uploaded files.}

\item{meddra_version}{A character or numeric. The MedDRA version
(e.g, "28.0", 28, "28.1", 28.1).}
}
\value{
A tibble with filtered MedDRA variables.
}
\description{
Prepare MedDRA data from uploaded files
}
