% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_volcano.R
\name{prepare_volcano}
\alias{prepare_volcano}
\title{Prepare adsl and adae for the volcano display}
\usage{
prepare_volcano(
  data,
  safety_variable = c("AEDECOD", "AEBODSYS", "MLG_label", "SMQ_NAME", "ocmq")
)
}
\arguments{
\item{data}{Results data.}

\item{safety_variable}{The hierarchy level to display: "AEDECOD", "AEBODSYS",
"MLG_label", "SMQ_NAME", "ocmq".}
}
\description{
Prepare adsl and adae for the volcano display
}
