% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{reorder_levels}
\alias{reorder_levels}
\title{Reorder safety variable levels to plot according to custom ordering}
\usage{
reorder_levels(data, variable, order_by, effect_measure, adjustment)
}
\arguments{
\item{data}{A results dataframe}

\item{variable}{A character, the lowest-level safety variable. "AEDECOD" or
"MLG_label"}

\item{order_by}{A character, either "effect" or "p-value"}

\item{effect_measure}{A character, either "RD" or "RR"}

\item{adjustment}{A character, either "FDR" or "DFDR"}
}
\value{
A dataframe, ordered by the selected method
}
\description{
Reorder safety variable levels to plot according to custom ordering
}
