% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{reorder_pattern}
\alias{reorder_pattern}
\title{Reorder a vector to show sensible defaults first, for text inputs (e.g.,
variable names)}
\usage{
reorder_pattern(vector, pattern)
}
\arguments{
\item{vector}{A character vector to reorder.}

\item{pattern}{A character, regex pattern to select sensible defaults.}
}
\value{
A reordered vector with elements containing the pattern in first
position.
}
\description{
Reorder a vector to show sensible defaults first, for text inputs (e.g.,
variable names)
}
