% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{row_half}
\alias{row_half}
\title{Shorthand for custom UI row - symmetrical halves}
\usage{
row_half(...)
}
\arguments{
\item{...}{Contents of the row.}
}
\description{
Shorthand for custom UI row - symmetrical halves
}
