% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{row_left_side}
\alias{row_left_side}
\title{Shorthand for custom UI row - left side}
\usage{
row_left_side(...)
}
\arguments{
\item{...}{Contents of the row.}
}
\description{
Shorthand for custom UI row - left side
}
