% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{row_right_side}
\alias{row_right_side}
\title{Shorthand for custom UI row - right side}
\usage{
row_right_side(...)
}
\arguments{
\item{...}{Contents of the row.}
}
\description{
Shorthand for custom UI row - right side
}
