% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{select_adae_data}
\alias{select_adae_data}
\title{Select adverse event data (used in DetectoR app)}
\usage{
select_adae_data(meddra_Version, mode, adae_file = NULL)
}
\arguments{
\item{meddra_Version}{Number of MedDRa Version}

\item{mode}{Different modes like SAS files, demo data, data from Server are available}

\item{adae_file}{Adverse event file (default:NULL)}
}
\value{
adae Transformed adverse event data set
}
\description{
Select adverse event data (used in DetectoR app)
}
