% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{select_adsl_data}
\alias{select_adsl_data}
\title{Select subject level data (used in DetectoR app)}
\usage{
select_adsl_data(mode, adsl_file = NULL)
}
\arguments{
\item{mode}{Different modes like SAS files, demo data, data from Server are available}

\item{adsl_file}{Subject level file (default:NULL)}
}
\value{
adae Transformed subject level data set
}
\description{
Select subject level data (used in DetectoR app)
}
