% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{show_error}
\alias{show_error}
\title{Show error message}
\usage{
show_error(error_message, label)
}
\arguments{
\item{error_message}{A character, the error message.}

\item{label}{A text to show before the error message.}
}
\description{
Show error message
}
