% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{study.adj.incidence}
\alias{study.adj.incidence}
\title{Calculate the study-size adjusted incidence rates}
\usage{
study.adj.incidence(split.data, alph = 95, strat_var = stratify_variable)
}
\arguments{
\item{split.data}{Data set with Counts.}

\item{alph}{Significance level.}

\item{strat_var}{Stratification variable.}
}
\value{
res A tibble with relative risk and risk differences.
}
\description{
Calculate the study-size adjusted incidence rates
}
