% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{study.strat.RR.and.RD}
\alias{study.strat.RR.and.RD}
\title{Calculate stratified relative risk and risk differences}
\usage{
study.strat.RR.and.RD(
  comb_data,
  variable,
  display,
  alpha = 0.95,
  stratify_variable
)
}
\arguments{
\item{comb_data}{adsl and adae data set combined}

\item{variable}{Description.}

\item{display}{Description.}

\item{alpha}{Description.}

\item{stratify_variable}{Stratification variable.}
}
\value{
res_new
}
\description{
Calculate stratified relative risk and risk differences
}
