% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{try_null}
\alias{try_null}
\title{tryCatch an expression, returns NULL if error}
\usage{
try_null(exp)
}
\arguments{
\item{exp}{An expression.}
}
\description{
tryCatch an expression, returns NULL if error
}
