% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_utils_upload.R
\name{update_values_input}
\alias{update_values_input}
\title{Update available values for variable selection}
\usage{
update_values_input(data_req, input_req, input_id, session)
}
\arguments{
\item{data_req}{Required dataset (usually, ADAE or ADSL).}

\item{input_req}{Required input (the corresponding variable).}

\item{input_id}{The ID of the input to update (character).}

\item{session}{The Shiny session.}
}
\description{
Update available values for variable selection
}
