% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_utils_upload.R
\name{update_variables_input}
\alias{update_variables_input}
\title{Update available variables for variable selection}
\usage{
update_variables_input(data_req, var_names, name_pattern, input_id, session)
}
\arguments{
\item{data_req}{Required dataset (usually, ADAE or ADSL).}

\item{var_names}{Vector with all column names of the corresponding dataset.}

\item{name_pattern}{A regexp pattern, with the most common names of the
variables. E.g., "SAFFL|SAFFN" for the safety population flag variable.}

\item{input_id}{The ID of the input to update (character).}

\item{session}{The Shiny session.}
}
\description{
Update available variables for variable selection
}
