% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_utils_upload.R
\name{upload_file}
\alias{upload_file}
\title{Upload valid file}
\usage{
upload_file(file)
}
\arguments{
\item{file}{A file uploaded from shiny::fileInput()}
}
\value{
A dataframe.
}
\description{
Upload valid file
}
