% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{validate_need}
\alias{validate_need}
\title{A shortcut for validate(need())}
\usage{
validate_need(object, label)
}
\arguments{
\item{object}{The object needed.}

\item{label}{A text to show when the object is not validated.}
}
\description{
A shortcut for validate(need())
}
