% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_data.R
\name{calculate_results}
\alias{calculate_results}
\title{Calculate all results for plotting}
\usage{
calculate_results(
  joint_data,
  adsl_filtered_data,
  variable = c("AEDECOD", "MLG_label", "SMQ_NAME", "ocmq", "AEBODSYS"),
  effect_measure = c("RR", "RD"),
  adjustment = c("FDR", "DFDR"),
  order_by = c("p-value", "effect"),
  study_strat = "None",
  alternative = c("two.sided", "less"),
  alpha = 0.05,
  filter = c("no_method", "one_percent", "min"),
  frequency_measure = c("proportions", "incidence rates"),
  duration_mode = c("none", "duration", "start_end_date"),
  ae_duration_variable = NULL,
  exposure_duration_variable = NULL,
  exposure_start_variable = NULL,
  exposure_end_variable = NULL,
  ae_start_variable = NULL
)
}
\arguments{
\item{joint_data}{Joined and filtered ADSL-ADAE data set with MedDRA data.}

\item{adsl_filtered_data}{A dataframe with ADSL after filtering.}

\item{variable}{A character, the variable name that will be shown in the plot
("AEDECOD", "MLG_label", "SMQ_NAME", "ocmq" or "AEBODSYS").}

\item{effect_measure}{Display option RR (relative risks) or RD (risk
differences).}

\item{adjustment}{Adjustment variable FDR or DFDR.}

\item{order_by}{Ordered by "p-value" or "effect"}

\item{study_strat}{A character, study stratification variable ("None" if
none).}

\item{alternative}{A character string describing the alternative hypothesis
(one or two sided).}

\item{alpha}{Alpha level of the test.}

\item{filter}{A character string if filter on adverse events should be
performed.}

\item{frequency_measure}{Measurement used (proportions or incidence rates).}

\item{duration_mode}{Way of calculating incidence risks denominators (time at
risk),
can be "duration", "start_end_date" or "none".}

\item{ae_duration_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_duration_variable}{Variables for calculation of time at risk
for incidence rates.}

\item{exposure_start_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{exposure_end_variable}{Variables for calculation of time at risk for
incidence rates.}

\item{ae_start_variable}{Variables for calculation of time at risk for
incidence rates.}
}
\value{
A dataframe with event counts, proportions/incidences, RR, RD,
 confidence intervals and p-values.
}
\description{
Calculate all results for plotting
}
