% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{count_ae_events}
\alias{count_ae_events}
\title{Count the AE events per treatment arm and AE category}
\usage{
count_ae_events(data, variable, overall = FALSE)
}
\arguments{
\item{data}{Combined ADAE-ADSL data}

\item{variable}{AE group category}

\item{overall}{Whether to calculate events per AE category, or overall}
}
\description{
Count the AE events per treatment arm and AE category
}
