% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{find_min_event}
\alias{find_min_event}
\title{Calculate the minimum number of events in verum}
\usage{
find_min_event(n1, n2, alternative = c("two.sided", "less"), alpha = 0.05)
}
\arguments{
\item{n1}{Number of patients in Verum}

\item{n2}{Number of patients in Comparator}

\item{alternative}{A character string describing the alternative hypothesis
(one or two sided)}

\item{alpha}{Alpha level of the test}
}
\value{
minimum number of events in verum
}
\description{
Calculate the minimum number of events in verum
}
