% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calculate_fct_count.R
\name{join_denominators}
\alias{join_denominators}
\title{Join denominators}
\usage{
join_denominators(data_events, data_denominators)
}
\arguments{
\item{data_events}{Count of events}

\item{data_denominators}{Count of denominators}
}
\description{
Join denominators
}
